python/requirements.txt
pyzmq==24.0.1
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1 